% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.Reg.R
\name{lin.Reg}
\alias{lin.Reg}
\title{Linear Regression}
\usage{
lin.Reg(Y, X, has_b0 = TRUE)
}
\arguments{
\item{Y}{A vector or a single column matrix response variable.}

\item{X}{A vector for univariate or matrix for multivariate of predictors.}

\item{has_b0}{A binary variable indicates whether an intercept is included.}
}
\value{
A list of attributes from the regression model.
}
\description{
This function implements the linear regression models.
It can work for both univariate linear regression, as well as multivariate linear regression.
}
