% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.Reg.R
\name{summary.lin.Reg}
\alias{summary.lin.Reg}
\title{Display the result of a linear regression model}
\usage{
\method{summary}{lin.Reg}(lmod)
}
\arguments{
\item{lmod}{The output object returned from lin.Reg}
}
\value{
displayed result of the fitted model
}
\description{
This function take the output of lin.Reg as input
then display the results in better organization,
includes beta coefficents, standard errors, t test statistics and p values
For model evaluation, the goodness of fit metric also shown, such as
R Squared, Adjusted R Squared, Mean Standard Errors, F test statistics, and p-value
}
